+++ vms/saevm/blocks/access.go
@@ -22,6 +22,14 @@ type (
 		Frontier
 		DB() ethdb.Database
 		XDB() types.ExecutionResults
+
+		// ResolvePendingToLastExecuted allows overriding of the default
+		// behaviour of [ResolveRPCNumber], which is to resolve
+		// [rpc.PendingBlockNumber] to the last-accepted block. This allows for
+		// compatibility with EVM-ecosystem tools that expect post-execution
+		// artefacts from the pending block, which is only possible in a
+		// synchronous paradigm.
+		ResolvePendingToLastExecuted() bool
 	}
 
 	// ConsensusCritical blocks are currently in use by a consensus mechanism,
@@ -87,7 +95,9 @@ var ErrNonCanonicalBlock = fmt.Errorf("%w: canonical block required", ErrFutureB
 // the corresponding block, treating named blocks as relative to execution:
 //
 //   - [rpc.PendingBlockNumber] is that returned by the [AcceptanceFrontier],
-//     its execution status being unknown but eventually guaranteed.
+//     its execution status being unknown but eventually guaranteed. However,
+//     if [Chain.ResolvePendingToLastExecuted] returns true, then the
+//     [ExecutionFrontier] is used instead. See said method for rationale.
 //   - [rpc.LatestBlockNumber] is that returned by the [ExecutionFrontier].
 //   - [rpc.SafeBlockNumber] and [rpc.FinalizedBlockNumber] are both that
 //     returned by the [SettlementFrontier].
@@ -111,16 +121,19 @@ var ErrNonCanonicalBlock = fmt.Errorf("%w: canonical block required", ErrFutureB
 // of the word. The "finalized" block is defined identically because we wish to
 // maintain monotonicity of the labels, and no further guarantees are possible
 // after settlement.
-func ResolveRPCNumber(f Frontier, bn rpc.BlockNumber) (uint64, error) {
-	tip := f.LastAccepted().Height()
+func ResolveRPCNumber(c Chain, bn rpc.BlockNumber) (uint64, error) {
+	if bn == rpc.PendingBlockNumber && c.ResolvePendingToLastExecuted() {
+		bn = rpc.LatestBlockNumber
+	}
+	tip := c.LastAccepted().Height()
 
 	switch bn {
 	case rpc.PendingBlockNumber:
 		return tip, nil
 	case rpc.LatestBlockNumber:
-		return f.LastExecuted().Height(), nil
+		return c.LastExecuted().Height(), nil
 	case rpc.SafeBlockNumber, rpc.FinalizedBlockNumber:
-		return f.LastSettled().Height(), nil
+		return c.LastSettled().Height(), nil
 	}
 
 	if bn < 0 {